+++ src/example/java/io/github/lucasstarsz/fastj/example/oldhelloworld/customscripts/PlayerScript.java
@@ -1,4 +1,4 @@
-package io.github.lucasstarsz.fastj.example.helloworld.customscripts;
+package io.github.lucasstarsz.fastj.example.oldhelloworld.customscripts;
 
 import io.github.lucasstarsz.fastj.math.Pointf;
 import io.github.lucasstarsz.fastj.graphics.game.GameObject;
@@ -7,7 +7,7 @@ import io.github.lucasstarsz.fastj.systems.behaviors.Behavior;
 import io.github.lucasstarsz.fastj.systems.input.keyboard.Keyboard;
 import io.github.lucasstarsz.fastj.systems.input.keyboard.Keys;
 
-import io.github.lucasstarsz.fastj.example.helloworld.scenes.GameScene;
+import io.github.lucasstarsz.fastj.example.oldhelloworld.scenes.GameScene;
 
 /** A custom script that moves the received game object using the WASD keys. */
 public class PlayerScript implements Behavior {
